refactor(tasks): move Go2 Arm shared base [skip ci]
