--- src/unilab/envs/locomotion/go2_arm/__init__.py
@@ -1,1 +1,0 @@
-"""Legacy Go2 Arm shared base pending migration to :mod:`unilab.tasks`."""
